split up shp methods into separate modules, add tf-test version
