+++ src/dolphin/phase_link/_mle_cpu.py
@@ -6,7 +6,7 @@ from dolphin._log import get_log
 from dolphin.utils import decimate
 from dolphin.workflows import ShpMethod
 
-from . import covariance, metrics, shp
+from . import covariance, metrics
 from .mle import mle_stack
 
 logger = get_log(__name__)
@@ -19,7 +19,7 @@ def run_cpu(
     beta: float = 0.01,
     reference_idx: int = 0,
     use_slc_amp: bool = True,
-    shp_method: str = ShpMethod.KL,
+    shp_method: str = ShpMethod.TF,
     avg_mag: Optional[np.ndarray] = None,
     var_mag: Optional[np.ndarray] = None,
     n_workers: int = 1,
@@ -45,8 +45,8 @@ def run_cpu(
         Whether to use the SLC amplitude when outputting the MLE estimate,
         or to set the SLC amplitude to 1.0. By default True.
     shp_method : Optional[str]
-        The SHP estimator to use. Either None, "KL" or "KS".
-        By default "KL", uses the KL distance estimator.
+        The SHP estimator to use.
+        By default "TF", uses a combination t-test/f-test
         If None, turns of the SHP search and uses a rectangular window.
     avg_mag : np.ndarray, optional
         The average magnitude of the SLC stack, used to find the SHP
@@ -67,15 +67,30 @@ def run_cpu(
     temp_coh : np.ndarray[np.float32]
         The temporal coherence at each pixel, shape (n_rows, n_cols)
     """
-    if shp_method.lower() == ShpMethod.KL:
+    halfwin_rowcol = (half_window["y"], half_window["x"])
+    # these two just use mean/var
+    if shp_method.lower() in (ShpMethod.KL, ShpMethod.TF):
         if avg_mag is None:
             avg_mag = np.mean(np.abs(slc_stack), axis=0)
         if var_mag is None:
             var_mag = np.var(np.abs(slc_stack), axis=0)
-        halfwin_rowcol = (half_window["y"], half_window["x"])
-        # TODO: threshold as a parameter?
+
+    if shp_method.lower() == ShpMethod.TF:
+        from . import _shp_tf_test
+
         logger.info("Estimating SHP neighbors using KL distance")
-        neighbor_arrays = shp.estimate_neighbors_kl_cpu(
+        neighbor_arrays = _shp_tf_test.estimate_neighbors(
+            avg_mag,
+            var_mag,
+            halfwin_rowcol=halfwin_rowcol,
+            n=slc_stack.shape[0],
+            alpha=0.05,  # TODO: make this a parameter
+        )
+    elif shp_method.lower() == ShpMethod.KL:
+        from . import _shp_kullback
+
+        logger.info("Estimating SHP neighbors using KL distance")
+        neighbor_arrays = _shp_kullback.estimate_neighbors_cpu(
             avg_mag, var_mag, halfwin_rowcol=halfwin_rowcol, threshold=0.5
         )
     elif shp_method == ShpMethod.RECT:
