+++ src/dolphin/phase_link/_utils.py
@@ -1,7 +1,7 @@
 import numba
-from numba import cuda
 
 
+@numba.njit
 def _get_slices(half_r: int, half_c: int, r: int, c: int, rows: int, cols: int):
     """Get the slices for the given pixel and half window size."""
     # Clamp min indexes to 0
@@ -11,8 +11,3 @@ def _get_slices(half_r: int, half_c: int, r: int, c: int, rows: int, cols: int):
     r_end = min(r + half_r + 1, rows)
     c_end = min(c + half_c + 1, cols)
     return (r_start, r_end), (c_start, c_end)
-
-
-# Make cpu and gpu compiled versions of the helper function
-_get_slices_cpu = numba.njit(_get_slices)
-_get_slices_gpu = cuda.jit(device=True)(_get_slices)
